refactor: Refactored path of page of login regular user
